refactor: organize dir on profile
