+++ src/app/favorite/service.rs
@@ -3,7 +3,7 @@ use crate::app::article::service::{fetch_article, FetchArticle};
 use crate::app::favorite::model::{CreateFavorite, DeleteFavorite, Favorite, FavoriteInfo};
 use crate::app::tag::model::Tag;
 use crate::app::user::model::User;
-use crate::appv2::features::profile::entities::profile::Profile;
+use crate::appv2::features::profile::entities::Profile;
 use crate::error::AppError;
 use diesel::pg::PgConnection;
