--- src/appv2/features/profile/adapters/mod.rs
@@ -1,2 +1,0 @@
-pub mod controllers;
-pub mod presenters;
