--- src/appv2/features/profile/domains/mod.rs
@@ -1,1 +1,0 @@
-pub mod profile_repository;
