--- src/appv2/features/profile/entities/mod.rs
@@ -1,1 +1,0 @@
-pub mod profile;
