+++ src/appv2/features/profile/mod.rs
@@ -1,4 +1,6 @@
-pub mod adapters;
-pub mod domains;
+pub mod controllers;
 pub mod entities;
+pub mod presenters;
+pub mod repositories;
+pub mod services;
 pub mod usecases;
