--- src/appv2/features/profile/usecases/mod.rs
@@ -1,2 +1,0 @@
-pub mod profile_usecase;
-pub mod services;
